Added basic edit and delete

+++ src/components/InfoModal.js
@@ -28,32 +28,49 @@ const style = {
   borderRadius: '10px',
 };
 
-export default function InfoModal({ openModal, setOpenModal, drawerState }) {
+export default function InfoModal({ openModal, setOpenModal, drawerState, userInfo }) {
 
   const userInfoArray = JSON.parse(localStorage.getItem("userInfo"));
 
-  const [name, setName] = useState(userInfoArray.length ? userInfoArray[0].name : '');
-  const [birthDate, setBirthDate] = useState(userInfoArray.length ? userInfoArray[0].birthDate : '');
-  const [gender, setGender] = useState(userInfoArray.length ? userInfoArray[0].gender : '');
-  const [mobile, setMobile] = useState(userInfoArray.length ? userInfoArray[0].setMobile : '');
+  const [name, setName] = useState(userInfo ? userInfo.name : '');
+  const [birthDate, setBirthDate] = useState(userInfo ? userInfo.birthDate : '');
+  const [gender, setGender] = useState(userInfo ? userInfo.gender : '');
+  const [mobile, setMobile] = useState(userInfo ? userInfo.mobile : '');
   const [showWarning, setShowWarning] = useState('');
-  
 
-  const saveInfo = () =>{
-    if(!name) setShowWarning('Please enter a valid name');
-    else if(!birthDate) setShowWarning('Please select a valid Birthdate');
-    else if(!gender) setShowWarning('Please select a valid gender');
-    else if(!mobile) setShowWarning('Please enter a valid name');
-    else{
+  useEffect(()=>{
+    setName(userInfo.name);
+    setBirthDate(userInfo.birthDate);
+    setGender(userInfo.gender);
+    setMobile(userInfo.mobile);
+  },[openModal])
+
+
+  const saveInfo = () => {
+    if (!name) setShowWarning('Please enter a valid name');
+    else if (!birthDate) setShowWarning('Please select a valid Birthdate');
+    else if (!gender) setShowWarning('Please select a valid gender');
+    else if (!mobile) setShowWarning('Please enter a valid name');
+    else {
+
       let obj = {
         name,
         gender,
         birthDate,
         mobile,
+        date: Date.now(),
       }
       let data = JSON.parse(localStorage.getItem('userInfo'))
       if (data) {
-        localStorage.setItem("userInfo", JSON.stringify([obj, ...data]));
+
+        const indexToReplace = data.findIndex(x => x.mobile === obj.mobile);
+        if (indexToReplace !== -1) {
+          data[indexToReplace] = obj;
+          localStorage.setItem("userInfo", JSON.stringify(data));
+        } else {
+          localStorage.setItem("userInfo", JSON.stringify([obj, ...data]));
+        }
+
       } else {
         localStorage.setItem("userInfo", JSON.stringify([obj]));
       }
@@ -73,36 +90,37 @@ export default function InfoModal({ openModal, setOpenModal, drawerState }) {
   //   }
   // }
 
-  useEffect(()=>setShowWarning(''),[name, gender, birthDate, mobile]);
+  useEffect(() => setShowWarning(''), [name, gender, birthDate, mobile]);
+  console.log('🚀___ ~ name, _______ :', name, gender, birthDate, mobile);
 
   return (
     <div>
-     {drawerState === 'Lushu' && <Modal
+      {drawerState === 'Lushu' && <Modal
         open={openModal}
         // onClose={setOpenModal}
         aria-labelledby="modal-modal-title"
         aria-describedby="modal-modal-description"
       >
         <Box sx={style}>
-        <div>
-          <p className='modal-title'>User Details</p>
-        </div>
+          <div>
+            <p className='modal-title'>User Details</p>
+          </div>
           <TextField
             required
             id="outlined-required"
             label="Name"
             fullWidth
             value={name}
-            onChange={(event)=>setName(event.target.value)}
-            sx={{marginBottom : '10px'}}
+            onChange={(event) => setName(event.target.value)}
+            sx={{ marginBottom: '10px' }}
           />
 
           <LocalizationProvider dateAdapter={AdapterDayjs} >
-            <DatePicker required format="DD/MM/YYYY" value={dayjs(birthDate)} onChange={(event)=>setBirthDate(dayjs(event))} sx={{marginBottom : '10px'}}/>
+            <DatePicker required format="DD/MM/YYYY" value={dayjs(birthDate)} onChange={(event) => setBirthDate(dayjs(event))} sx={{ marginBottom: '10px' }} />
           </LocalizationProvider>
 
 
-          <FormControl sx={{marginBottom : '10px'}} fullWidth>
+          <FormControl sx={{ marginBottom: '10px' }} fullWidth>
             <InputLabel id="demo-simple-select-label">Gender</InputLabel>
             <Select
               required
@@ -110,7 +128,7 @@ export default function InfoModal({ openModal, setOpenModal, drawerState }) {
               id="demo-simple-select"
               value={gender}
               label="Gender"
-              onChange={(event)=>setGender(event.target.value)}
+              onChange={(event) => setGender(event.target.value)}
             >
               <MenuItem value={'male'}>Male</MenuItem>
               <MenuItem value={'female'}>Female</MenuItem>
@@ -125,11 +143,14 @@ export default function InfoModal({ openModal, setOpenModal, drawerState }) {
             fullWidth
             value={mobile}
             type='number'
-            onChange={(event)=>setMobile(event.target.value)}
-            sx={{marginBottom : '10px'}}
+            onChange={(event) => setMobile(event.target.value)}
+            sx={{ marginBottom: '10px' }}
           />
 
-          <Button variant="contained"  color="success" sx={{marginTop:'10px', padding : '10px 24px', marginLeft : 'auto', display: 'block'}} onClick={saveInfo}>Save</Button>
+          <div className='button-allign'>
+            {userInfoArray.length && <Button variant="contained" color="warning" sx={{ marginTop: '10px', padding: '10px 24px' }} onClick={() => setOpenModal(false)}>Cancel</Button>}
+            <Button variant="contained" color="success" sx={{ marginTop: '10px', padding: '10px 24px' }} onClick={saveInfo}>Save</Button>
+          </div>
         </Box>
       </Modal>}
